- allowed items hp and mp values to go over 255 (since most modern servers have hp/mp pots that fill more then 255) - taken out the code that allowed higher levels regen faster from pots (you still regen more / level but not super fast like before)

+++ Server/ItemInfoForm.cs
@@ -764,9 +764,9 @@ namespace Server
         {
             if (ActiveControl != sender) return;
 
-            byte temp;
+            ushort temp;
 
-            if (!byte.TryParse(ActiveControl.Text, out temp))
+            if (!ushort.TryParse(ActiveControl.Text, out temp))
             {
                 ActiveControl.BackColor = Color.Red;
                 return;
@@ -782,9 +782,9 @@ namespace Server
 
             if (ActiveControl != sender) return;
 
-            byte temp;
+            ushort temp;
 
-            if (!byte.TryParse(ActiveControl.Text, out temp))
+            if (!ushort.TryParse(ActiveControl.Text, out temp))
             {
                 ActiveControl.BackColor = Color.Red;
                 return;
